drivers: input: touchscreen: vdd_ana regulator fix
Due to android resume after boot, touch driver was
getting DRM notifier for resume and enabling the
regulators which was resulting in incrementing
reference count.
Since regulators were getting enabled during probe
additional resume without a corresponding suspend
was resulting in additional use count, keeping the
regulator ON which was causing additional current.
This change is fixing the issue by enabling the
regulator only if regulator was not enabled.
Also added a delay for regulators to quiesce or
ramp up properly.

Change-Id: Iea4cb26f1bcaa5f5111fba3a66d18399e4ac0ff9
Signed-off-by: Raj Pandey <quic_rajpande@quicinc.com>